--- scripts/game.js
@@ -1,231 +1,0 @@
-var nativeWidth = 320;
-var nativeHeight = 180;
-var platforms, players, text, sfx;
-
-var resize = function resize() {
-  document.body.style.zoom = window.innerWidth / nativeWidth;
-};
-
-var checkForGameOver = function checkForGameOver() {
-  var alivePlayers = [];
-  players.children.forEach(function(player) {
-    if (!player.isDead) {
-      alivePlayers.push(player.name);
-    }
-  });
-  if (alivePlayers.length === 1) {
-    text.setText(alivePlayers[0] + '  wins!\nClick  to  restart');
-    text.visible = true;
-    game.input.onDown.addOnce(restart, this); // restart game on mouse click
-  }
-};
-
-var preload = function preload() {
-  var utils = require('./utils.js');
-  sfx = require('./sfx.js');
-
-  resize();
-  window.onresize = utils.debounce(resize, 100);
-
-  game.load.image('pink', 'images/pink.png');
-  game.load.image('yellow', 'images/yellow.png');
-  game.load.image('blue', 'images/blue.png');
-  game.load.image('orange', 'images/orange.png');
-  game.load.image('purple', 'images/purple.png');
-  game.load.image('green', 'images/green.png');
-  game.load.image('white', 'images/white.png');
-  game.load.spritesheet('hearts', 'images/hearts.png', 9, 5); // player health
-};
-
-var create = function create() {
-  game.physics.startSystem(Phaser.Physics.ARCADE);
-  game.world.setBounds(0, -nativeHeight, nativeWidth, nativeHeight * 3); // allow anything as tall as world to fall off-screen up or down
-
-  var buildPlatforms = require('./map.js');
-  platforms = buildPlatforms(game);
-
-  game.input.gamepad.start();
-
-  // TODO: why is this font still anti-aliased?
-  var fontStyle = { font: "12px Hellovetica", fill: "#eee", align: "center", boundsAlignH: "center", boundsAlignV: "middle" };
-  text = game.add.text(0, 0, '', fontStyle);
-  text.setTextBounds(0, 0, nativeWidth, nativeHeight);
-
-  players = game.add.group();
-  restart();
-};
-
-var restart = function() {
-  text.visible = false;
-
-  while (players.children.length > 0) {
-    players.children[0].destroy();
-  }
-
-  var createPlayer = require('./player.js');
-
-  var player1 = {
-    name: 'Blue',
-    color: 'blue',
-    gamepad: game.input.gamepad.pad1,
-    position: {
-      x: 72, y: 44
-    },
-  };
-
-  var player2 = {
-    name: 'Yellow',
-    color: 'yellow',
-    gamepad: game.input.gamepad.pad2,
-    position: {
-      x: 248, y: 44
-    },
-    orientation: 'left',
-  };
-
-  var player3 = {
-    name: 'Green',
-    color: 'green',
-    gamepad: game.input.gamepad.pad3,
-    keys: {
-      up: 'W', down: 'S', left: 'A', right: 'D', attack: 'Q'
-    },
-    position: {
-      x: 72, y: 136
-    },
-  };
-
-  var player4 = {
-    name: 'Purple',
-    color: 'purple',
-    gamepad: game.input.gamepad.pad4,
-    keys: {
-      up: 'I', down: 'K', left: 'J', right: 'L', attack: 'U'
-    },
-    position: {
-      x: 248, y: 136
-    },
-    orientation: 'left',
-  };
-
-  players.add(createPlayer(game, player1));
-  players.add(createPlayer(game, player2));
-  players.add(createPlayer(game, player3));
-  players.add(createPlayer(game, player4));
-};
-
-var update = function update() {
-  game.physics.arcade.collide(players, platforms);
-  // TODO: how do i do this on the player itself without access to players? or should i add a ftn to player and set that as the cb?
-  game.physics.arcade.collide(players, players, function handlePlayerCollision(playerA, playerB) {
-     /* let's not knock anybody around if something's on one of these dudes'/dudettes' heads.
-     prevents cannonball attacks and the like, and allows standing on heads.
-     note: still need to collide in order to test touching.up, so don't move this to allowPlayerCollision! */
-    if (playerA.body.touching.up || playerB.body.touching.up) {
-      return;
-    }
-
-    function temporarilyDisableCollision(player) {
-      player.isCollidable = false;
-      setTimeout(function() {
-        player.isCollidable = true;
-      }, 100);
-    }
-
-    function bounce() {
-      sfx.bounce();
-
-      var bounceVelocity = 100;
-      var velocityA = velocityB = bounceVelocity;
-      if (playerA.position.x > playerB.position.x) {
-        velocityB *= -1;
-      } else {
-        velocityA *= -1;
-      }
-      playerA.body.velocity.x = velocityA;
-      playerB.body.velocity.x = velocityB;
-      playerA.isRolling = false;
-      playerB.isRolling = false;
-    }
-
-    function fling() {
-      sfx.bounce();
-
-      var playerToFling;
-      var playerToLeave;
-      if (playerA.isDucking) {
-        playerToFling = playerB;
-        playerToLeave = playerA;
-      } else {
-        playerToFling = playerA;
-        playerToLeave = playerB;
-      }
-      temporarilyDisableCollision(playerToFling);
-      var flingXVelocity = 150;
-      if (playerToFling.position.x > playerToLeave.position.x) {
-        flingXVelocity *= -1;
-      }
-      playerToFling.body.velocity.x = flingXVelocity;
-      playerToFling.body.velocity.y = -150;
-    }
-
-    function pop() {
-      sfx.bounce();
-
-      var playerToPop;
-      if (playerA.isRolling) {
-        playerToPop = playerB;
-      } else {
-        playerToPop = playerA;
-      }
-      temporarilyDisableCollision(playerToPop);
-      playerToPop.body.velocity.y = -150;
-    }
-
-    var bothRolling = playerA.isRolling && playerB.isRolling;
-    var bothStanding = !playerA.isDucking && !playerB.isDucking;
-    var neitherRolling = !playerA.isRolling && !playerB.isRolling;
-    var eitherDucking = playerA.isDucking || playerB.isDucking;
-    var eitherRunning = Math.abs(playerA.body.velocity.x) > 28 || Math.abs(playerB.body.velocity.x) >= 28;
-    var eitherRolling = playerA.isRolling || playerB.isRolling;
-    var eitherStanding = !playerA.isDucking || !playerB.isDucking;
-
-    switch (true) {
-      case bothRolling || bothStanding:
-        bounce();
-        break;
-      case neitherRolling && eitherRunning && eitherDucking:
-        fling();
-        break;
-      case eitherRolling && eitherStanding:
-        pop();
-        break;
-    }
-
-    // if only one of the touching players is attacking...
-    if (playerA.isAttacking !== playerB.isAttacking) {
-      var victim = playerA.isAttacking ? playerB : playerA;
-      if (playerA.orientation !== playerB.orientation) {
-        victim.actions.takeDamage(1);
-      } else {
-        victim.actions.takeDamage(2); // attacked from behind for double damage
-      }
-    }
-
-  }, function allowPlayerCollision(playerA, playerB) {
-    // don't allow collision if either player isn't collidable.
-    // also disallow if player is in limbo below the screen :]
-    if (!playerA.isCollidable || !playerB.isCollidable || playerA.position.y > nativeHeight || playerB.position.y > nativeHeight) {
-      return false;
-    }
-    return true;
-  });
-};
-
-var game = new Phaser.Game(nativeWidth, nativeHeight, Phaser.AUTO, 'game', {
-  preload: preload,
-  create: create,
-  update: update,
-}, false, false); // disable anti-aliasing
-
-module.exports = checkForGameOver;
