+++ scripts/states/credits.js
@@ -1,0 +1,2 @@
+var credits = {
+};
